commonize updateStateAfterSave
